refactor: merge duplicates
